Lazy creation of basic objects
When objects are requested from $app and passed to `AnnotationsServiceProvider`, they are instantiated.

This is a bit wasteful, but more importantly some of them can only be built from the cli and not in php-fpm, as our configuration disables the `pcntl_*` functions for security reasons.

We should probably go for a similar pattern for the other objects as well, but they don't break the application on our servers.